initial support for packaging for chrome/android/ios
Using https://github.com/MobileChromeApps/mobile-chrome-apps to convert the chrome app into a cordova app. This (in theory) gives us the potential to code to the Chrome App API and have that API translate over to Cordova for us. The status of the API can be found here:

- https://github.com/MobileChromeApps/mobile-chrome-apps/blob/master/docs/APIStatus.md

A blog post introducing this approach of "mobilizing chrome apps" can be found here:

- http://blog.chromium.org/2014/01/run-chrome-apps-on-mobile-using-apache.html

See README.md for more details.